Clean up Lance writer utilities
